[IMP] runbot: add a job_type on the build too
A job_type was just added to the branch. Sometimes, some builds have to
be built without running despite the branch configuration.

For example, the master branch can be configured to create builds with
testing and running but nightly multiple builds can be generated with
testing only.

For this purpose, this commit adds a computed field job_type on the
build object. By default, this field inherits the value from the build's
branch but it can be overwritten to something else.